func-test: separate to three different tests
