+++ bin/strata-client/src/rpc_server.rs
@@ -12,6 +12,8 @@ use futures::TryFutureExt;
 use jsonrpsee::core::RpcResult;
 use strata_bridge_relay::relayer::RelayerHandle;
 use strata_btcio::{broadcaster::L1BroadcastHandle, writer::EnvelopeHandle};
+#[cfg(feature = "debug-utils")]
+use strata_common::bail_manager::BAIL_SENDER;
 use strata_consensus_logic::{
     checkpoint::CheckpointHandle, csm::state_tracker::reconstruct_state, l1_handler::verify_proof,
     sync_manager::SyncManager,
@@ -53,8 +55,6 @@ use strata_storage::L2BlockManager;
 use strata_zkvm::ProofReceipt;
 use tokio::sync::{oneshot, Mutex};
 use tracing::*;
-#[cfg(feature = "debug-utils")]
-use {strata_common::bail_manager::BAIL_SENDER, strata_rpc_api::StrataDebugApiServer};
 
 use crate::extractor::{extract_deposit_requests, extract_withdrawal_infos};
 
@@ -883,7 +883,7 @@ impl<D: Database + Sync + Send + 'static> StrataDebugApiServer for StrataDebugRp
 
     async fn set_bail_context(&self, _ctx: String) -> RpcResult<()> {
         #[cfg(feature = "debug-utils")]
-        let _sender = BAIL_SENDER.send(Some(ctx));
+        let _sender = BAIL_SENDER.send(Some(_ctx));
         Ok(())
     }
 }
