+++ crates/rpc/api/src/lib.rs
@@ -191,6 +191,6 @@ pub trait StrataDebugApi {
     #[method(name = "debug_getClientStateAtIdx")]
     async fn get_clientstate_at_idx(&self, idx: u64) -> RpcResult<Option<ClientState>>;
     /// for exiting the client based on context
-    #[method(name = "stratadebug_bail")]
+    #[method(name = "debug_bail")]
     async fn set_bail_context(&self, ctx: String) -> RpcResult<()>;
 }
