+++ functional-tests/utils/utils.py
@@ -579,3 +579,58 @@ def confirm_btc_withdrawal(
     assert (
         btc_balance == original_balance + expected_increase
     ), "BTC balance after withdrawal is not as expected"
+
+
+def handle_bailout(seq, seqrpc, bail_context):
+    """
+    Handles the bailout process for the given sequencer RPC.
+
+    Args:
+        seqrpc: The RPC interface for the sequencer.
+        seq: The sequencer service instance.
+        bail_context: The context in which to trigger the bailout.
+        cur_chain_tip: The current chain tip slot to monitor progress.
+
+    Raises:
+        AssertionError: If the bailout or chain tip progress fails.
+    """
+
+    def client_status(seqrpc, cur_chain_tip):
+        def another():
+            d = seqrpc.strata_clientStatus()
+            print(d)
+            return d["chain_tip_slot"] > cur_chain_tip
+
+        return another
+
+    # wait for 2 seconds for chain tip slot to accumulate.
+    # Since the chain tip requirement is not exact, we can sleep here
+    time.sleep(2)
+    cur_chain_tip = seqrpc.strata_clientStatus()["chain_tip_slot"]
+    print(cur_chain_tip)
+
+    # Trigger the bailout
+    seqrpc.debug_bail(bail_context)
+
+    # Ensure the sequencer bails out
+    wait_until(
+        lambda: check_sequencer_down(seqrpc),
+        error_with="Sequencer didn't bail out",
+    )
+    # Stop the sequencer to update bookkeeping, we know the sequencer has
+    # already stopped
+    seq.stop()
+
+    # Restart the sequencer
+    seq.start()
+
+    wait_until(
+        lambda: not check_sequencer_down(seqrpc), error_with="Sequencer didn't start", timeout=20
+    )
+
+    # Ensure the chain tip progresses
+    wait_until(
+        client_status(seqrpc, cur_chain_tip),
+        error_with="chain tip slot not progressing",
+        timeout=20,
+    )
